move from time -> clock
